Properly identify test modules
